Finished sql datastore refactoring
