+++ Makefile
@@ -11,7 +11,7 @@ dep: ## Get the dependencies
 	@go mod download
 
 lint: ## Lint Golang files
-	@golint -set_exit_status ${PKG_LIST}
+	@golangci-lint -c .golangci.yml run
 
 vet: ## Run go vet
 	@go vet ${PKG_LIST}
