+++ internal/pkg/data/factory.go
@@ -6,11 +6,11 @@ import (
 	"github.com/Foundato/kelon/pkg/data"
 )
 
-func MakeDatastores(config *configs.DatastoreConfig) map[string]*data.Datastore {
-	result := make(map[string]*data.Datastore)
+func MakeDatastores(config *configs.DatastoreConfig) map[string]*data.DatastoreTranslator {
+	result := make(map[string]*data.DatastoreTranslator)
 	for dsName, ds := range config.Datastores {
 		if ds.Type == data.TypeMysql || ds.Type == data.TypePostgres {
-			newDs := NewSQLDatastore()
+			newDs := NewSQLDatastore(NewSQLDatastoreExecutor())
 			logging.LogForComponent("factory").Infof("Init SqlDatastore of type [%s] with alias [%s]", ds.Type, dsName)
 			result[dsName] = &newDs
 			continue
